requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

backports.entry-points-selectable==1.1.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.2
filelock==3.0.12
Flask==2.2.3
fonttools==4.38.0
huggingface-hub==0.12.1
idna==3.3
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langdetect==1.0.9
MarkupSafe==2.1.2
matplotlib==3.7.0
nltk==3.8.1
numpy==1.24.2
oauthlib==3.1.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
platformdirs==2.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.26.0
requests-oauthlib==1.3.0
scipy==1.10.1
six==1.16.0
soupsieve==2.4
tokenizers==0.13.2
https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp39-cp39-linux_x86_64.whl
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.64.1
transformers==4.26.1
tweepy==4.4.0
typing_extensions==4.5.0
urllib3==1.26.7
virtualenv==20.7.0
Werkzeug==2.2.3
wordcloud==1.8.2.2
zipp==3.14.0
